test(cli): clean up relation tests
- rename files to follow the pattern `relation.{type}.integration.ts`
- move relation-type-specific tests from `relation.integration.ts`
  to relation-type-specific test files
